pyproject.toml
[project]
name = "aethergyre"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "alembic>=1.18.2",
    "fastapi>=0.128.0",
    "ijson>=3.4.0.post0",
    "psycopg2>=2.9.11",
    "pydantic-settings>=2.13.1",
    "sqlalchemy>=2.0.46",
    "uvicorn>=0.41.0",
]
